split out markdown generation
